Added link in developing with dev tools
I don't think you'll merge this PR because it looks like it's just promoting a commercial product.

I made a 'remote browser isolation' platform and share this free demo of it here. 

I heavily rely on Chrome devtools for this task so want to give back to the community by showing what's possible.

As I'm unwilling to release the source (because the full product is money-making) I'm quite sure you'll be unable to include it here.